Fix two bugs found in first real cluster test
1. matlab_tm_wrapper.m: remove_outliers was set via
   ~strcmp(getenv('REMOVE_OUTLIERS'), '0'), which produces a MATLAB
   logical, not a numeric double. Downstream, cifti_conn_matrix_for_
   wrapper_continous's make_numeric() helper checks isnumeric() before
   deciding whether to str2num() the value -- and isnumeric(true) is
   false in MATLAB, so it fell into the str2num branch on a logical,
   which errors immediately ("Input must be a character vector or
   string scalar"). Wrapped in double() to match the original
   hardcoded `remove_outliers = 1` numeric type.

2. motion_and_dtseries_bagging_cleaned.m: all 8 shuffle-order cache
   files (one per --shuffle option) had a check-then-act race --
   `if exist(path), load(path), else <build + save(path)>`. Since
   orchestrate_PCM.sh submits every permutation in a run back-to-back,
   the very first run for a given subject/session/truncation can have
   two or more permutations simultaneously see the cache missing and
   race to build it, and one process's load() can catch another
   mid-save(), reading a truncated .mat ("Not a binary MAT-file").
   Added an atomic_save() helper (write to a unique temp file, then
   movefile -- atomic rename on the same filesystem) and switched all
   8 save() call sites to use it. A concurrent reader now either sees
   the cache missing or fully written, never partial.

Both found from a real --num 2 --startmins 70 matlab_tm test run on
the cluster -- perm 1 hit bug #1, perm 2 hit bug #2 independently.

+++ template_matching/matlab_tm_wrapper.m
@@ -42,7 +42,7 @@ overlap_method = 'smooth_then_derivative';
 % REMOVE_OUTLIERS=0 (run_PCM.sh's --remove-outliers flag) to skip the
 % standard-deviation-based outlier removal cifti_conn_matrix_for_wrapper_continous
 % does on top of FD-based motion censoring.
-remove_outliers = ~strcmp(getenv('REMOVE_OUTLIERS'), '0');
+remove_outliers = double(~strcmp(getenv('REMOVE_OUTLIERS'), '0'));
 additional_mask = 'none';
 
 %check input format
